Check for valid result from getpwnam
Previously, trying to generate a key for a user that does not exist on
the master would result in an error because the `$userhash` was undef
and not a hash.

Actually, now that I think of it, this might be quite correct and is one
of the areas where rspec-puppet testing fails, because mocking a Puppet
function to return (Ruby) "undef" might not be quite the same as
(Puppet function) `getpwnam` returning ":undef". (Which might also be
something I need to address with the `name_service_lookups` module.)